python_agent_service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.1
langchain-openai==0.0.5 